requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
